requirements.txt
Flask
pyserial
# dev/test dependencies are kept in dev-requirements.txt
